+++ pyqmix_backend/app.py
@@ -66,6 +66,10 @@ initiate_or_disconnect_pumps = api.model('Initiate pumps', {
                             required=True,
                             example=True)})
 
+stop_pumps = api.model('Stop pumps', {
+    'stop': Boolean(description='Stop pumps',
+                    required=True,
+                    example=True)})
 
 ## --- Endpoints --- ##
 
@@ -90,11 +94,14 @@ class SetUpConfig(Resource):
 
     def get(self):
         # Return a list of available config-dirs and send to frontend
-
         list_of_available_configurations = config.get_available_qmix_configs()
         list_of_available_syringe_sizes = list(syringes.keys())
 
-        setup_dict = {'available_configs': list_of_available_configurations,
+        # Return status of pump-setup
+        config_setup = is_config_set_up()
+
+        setup_dict = {'is_config_set_up': config_setup,
+                      'available_configs': list_of_available_configurations,
                       'available_syringes': list_of_available_syringe_sizes}
         return setup_dict
 
@@ -121,10 +128,7 @@ class InitiateOrDisconnectPumps(Resource):
             pump_state = get_pump_state(pump_id)
             pump_states.append(pump_state)
 
-        # Return status of pump-setup
-        config_setup = is_config_set_up()
-
-        system_state = {'config_setup': config_setup, 'pump_states': pump_states}
+        system_state = {'pump_states': pump_states}
 
         return system_state
 
@@ -142,6 +146,14 @@ class InitiateOrDisconnectPumps(Resource):
 
         return status
 
+@api.route('/api/stop')
+class StopPumps(Resource):
+
+    @api.expect(stop_pumps)
+    def put(self):
+        if not app.config['test_session']:
+            list(session_paramters['pumps'].values())[0].stop_all_pumps()
+
 
 @api.route('/api/pumps/<int:pump_id>')
 class Pumps(Resource):
@@ -155,6 +167,8 @@ class Pumps(Resource):
     def put(self, pump_id):
         payload = request.json
         action = payload['action']
+
+        print('action: ' + action)
 
         if action == 'referenceMove':
             pump_reference_move(pump_id)
@@ -172,7 +186,7 @@ class Pumps(Resource):
 def is_config_set_up():
 
     if app.config['test_session']:
-        if session_paramters['get_pumps_states_call_count'] < 2:
+        if session_paramters['get_pumps_states_call_count'] < 1:
             return False
         else:
             return True