+++ CHANGES.md
@@ -1,3 +1,9 @@
+2018.10.26
+----------
+- Abort current pump operation if user initiates a new one
+- User can terminate current pump routine, but not modify it   
+- Only read pyqmix.config during startup
+
 2018.10.21
 ----------
 - Remove superfluous files